confirm delete product with alert

+++ src/components/Product.js
@@ -2,12 +2,25 @@ import React from 'react';
 import { Link } from 'react-router-dom';
 import { useDispatch } from 'react-redux';
 import { deleteProductsAction } from './../actions/productActions'
+import Swal from 'sweetalert2'
 
 const Product = ({product}) => {
   const dispatch = useDispatch()
 
   const confirmDelete = id => {
-    dispatch(deleteProductsAction(id))
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+      if (result.value) {
+        dispatch(deleteProductsAction(id)) // Call to action delete
+      }
+    })
   }
 
   return (
